# Reduce allocations when traversing a database