--- a/cursor.go
+++ b/cursor.go
@@ -36,12 +36,12 @@ func (c *Cursor) Close() {
 // Managing our own free list of stacks seems to work much better than using a
 // sync.Pool - potentially at the expense of a little higher data usage if
 // the application uses a lot of cursors at once
-type FreeStack struct {
+type freeStack struct {
 	sync.Mutex
 	items [][]elemRef
 }
 
-func (fs *FreeStack) Get() []elemRef {
+func (fs *freeStack) Get() []elemRef {
 	fs.Lock()
 	l := len(fs.items)
 	var s []elemRef
@@ -53,21 +53,21 @@ func (fs *FreeStack) Get() []elemRef {
 	return s
 }
 
-func (fs *FreeStack) Put(e []elemRef) {
+func (fs *freeStack) Put(e []elemRef) {
 	e = e[:0]
 	fs.Lock()
 	fs.items = append(fs.items, e)
 	fs.Unlock()
 }
 
-var stackPool = FreeStack{}
+var stackPool = freeStack{}
 
-type FreeCursors struct {
+type freeCursors struct {
 	sync.Mutex
 	items []*Cursor
 }
 
-func (fs *FreeCursors) Get() *Cursor {
+func (fs *freeCursors) Get() *Cursor {
 	fs.Lock()
 	l := len(fs.items)
 	var c *Cursor
@@ -81,14 +81,14 @@ func (fs *FreeCursors) Get() *Cursor {
 	return c
 }
 
-func (fs *FreeCursors) Put(c *Cursor) {
+func (fs *freeCursors) Put(c *Cursor) {
 	c.bucket = nil
 	fs.Lock()
 	fs.items = append(fs.items, c)
 	fs.Unlock()
 }
 
-var cursorPool = FreeCursors{}
+var cursorPool = freeCursors{}
 
 // Bucket returns the bucket that this cursor was created from.
 func (c *Cursor) Bucket() *Bucket {