@philpearl 2016-05-13
Improving the other test cases was just a case of writing the change properly. Here's the latest benchmarks
```
$ go test -bench . -benchmem
seed: 81596
quick settings: count=5, items=1000, ksize=1024, vsize=1024
00010000000000000000000000000000
PASS
Benchmark_FreelistRelease10K-8   	   10000	    193359 ns/op	   82501 B/op	       5 allocs/op
Benchmark_FreelistRelease100K-8  	    1000	   1784276 ns/op	  805266 B/op	       5 allocs/op
Benchmark_FreelistRelease1000K-8 	     300	   4970484 ns/op	 8049008 B/op	       5 allocs/op
Benchmark_FreelistRelease10000K-8	      30	  41784659 ns/op	80413040 B/op	       5 allocs/op
BenchmarkDBBatchAutomatic-8      	     500	   2741149 ns/op	  347806 B/op	    8174 allocs/op
BenchmarkDBBatchSingle-8         	      10	 165493580 ns/op	16899824 B/op	   53639 allocs/op
BenchmarkDBBatchManual10x100-8   	     300	   4216260 ns/op	  838531 B/op	    3986 allocs/op
BenchmarkDBTraverse-8            	 1000000	      1503 ns/op	      64 B/op	       1 allocs/op
ok  	github.com/boltdb/bolt	154.254s
```