@philpearl 2016-05-16
Added free pools for buckets. This only really helps for read transactions, and only if the application cooperates and says when it is done with the bucket. In my case I now have an average of 0 allocations per operation when scanning my database - a huge win.

```
$ go test -bench . -benchmem
seed: 30215
quick settings: count=5, items=1000, ksize=1024, vsize=1024
00010000000000000000000000000000
PASS
Benchmark_FreelistRelease10K-8   	   10000	    176919 ns/op	   82501 B/op	       5 allocs/op
Benchmark_FreelistRelease100K-8  	    1000	   2129799 ns/op	  805260 B/op	       5 allocs/op
Benchmark_FreelistRelease1000K-8 	     300	   5120436 ns/op	 8049008 B/op	       5 allocs/op
Benchmark_FreelistRelease10000K-8	      30	  46394099 ns/op	80413040 B/op	       5 allocs/op
BenchmarkDBBatchAutomatic-8      	     500	   2983732 ns/op	  350524 B/op	    8213 allocs/op
BenchmarkDBBatchSingle-8         	      10	 168321261 ns/op	16700965 B/op	   53583 allocs/op
BenchmarkDBBatchManual10x100-8   	     300	   4347885 ns/op	  835096 B/op	    3987 allocs/op
BenchmarkDBTraverse-8            	 1000000	      1471 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/boltdb/bolt	160.963s
```