@benbjohnson 2016-09-16
@philpearl Just to give you an update, I've been testing out the code and it seems to work fine but I'm concerned about the overhead of locks within `sync.Pool` for users who don't utilize `Close()`. 

Since `Bucket` objects are not safe for concurrent use, I wonder if a simpler approach would be to have an embedded `Cursor` & `[]elemRef` inside `Bucket` that it can use for all its internal operations (`Get()`, Put()`, etc) and only incur one `Bucket` allocation instead of an allocation on every key access.

That doesn't solve the `Bucket` allocation issue though. I think that's a less frequent issue though. One option would be to optionally attach the `bucketPool` to the `Tx`. Otherwise it would always return new allocations.