@josharian 2016-12-18
What's the status here? Embedding definitely sounds like a better approach than sync.Pool in this case.