@nuss-justin 2016-12-18
I just ran the benchmarks with an embedded cursor in bucket for internal use, to see how much improves the performance.

See [this gist](https://gist.github.com/nuss-justin/b8a57b2370ae6504ebfbed32bb8eb274) for the patch. I haven't really looked at the code so it's probably missing something, but it passes all tests with and without race detector.

This is what I got:

```
name                      old time/op    new time/op    delta
_FreelistRelease10K-8       16.7µs ± 1%    16.5µs ± 1%   -1.21%  (p=0.000 n=10+10)
_FreelistRelease100K-8       213µs ± 1%     212µs ± 1%   -0.31%  (p=0.034 n=10+10)
_FreelistRelease1000K-8     3.30ms ± 0%    3.28ms ± 1%   -0.65%  (p=0.028 n=9+10)
_FreelistRelease10000K-8    31.2ms ± 1%    30.4ms ± 1%   -2.65%  (p=0.000 n=9+9)
DBBatchAutomatic-8          1.71ms ± 1%    1.48ms ± 0%  -13.07%  (p=0.000 n=9+10)
DBBatchSingle-8             27.4ms ± 1%    27.4ms ± 0%     ~     (p=0.280 n=10+10)
DBBatchManual10x100-8       1.67ms ± 1%    1.48ms ± 1%  -11.57%  (p=0.000 n=9+10)

name                      old alloc/op   new alloc/op   delta
_FreelistRelease10K-8       82.5kB ± 0%    82.5kB ± 0%     ~     (all equal)
_FreelistRelease100K-8       805kB ± 0%     805kB ± 0%     ~     (all equal)
_FreelistRelease1000K-8     8.05MB ± 0%    8.05MB ± 0%     ~     (all equal)
_FreelistRelease10000K-8    80.4MB ± 0%    80.4MB ± 0%     ~     (all equal)
DBBatchAutomatic-8           384kB ± 0%     302kB ± 0%  -21.51%  (p=0.000 n=9+10)
DBBatchSingle-8             16.2MB ± 2%    16.2MB ± 1%     ~     (p=0.481 n=10+10)
DBBatchManual10x100-8        901kB ± 1%     819kB ± 2%   -9.09%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
_FreelistRelease10K-8         5.00 ± 0%      5.00 ± 0%     ~     (all equal)
_FreelistRelease100K-8        5.00 ± 0%      5.00 ± 0%     ~     (all equal)
_FreelistRelease1000K-8       5.00 ± 0%      5.00 ± 0%     ~     (all equal)
_FreelistRelease10000K-8      5.00 ± 0%      5.00 ± 0%     ~     (all equal)
DBBatchAutomatic-8           10.2k ± 0%      8.2k ± 0%  -19.77%  (p=0.000 n=10+10)
DBBatchSingle-8              58.8k ± 0%     57.8k ± 0%   -1.69%  (p=0.000 n=10+9)
DBBatchManual10x100-8        6.03k ± 1%     4.04k ± 1%  -33.00%  (p=0.000 n=10+10)
```

go version devel +61db2e4efa Fri Dec 16 17:03:25 2016 +0000 linux/amd64