@josharian 2016-12-19
@nuss-justin I don't know the codebase well enough to review. However, I suspect that you want to zero the embedded Cursor (or at least some of its fields) when it is not needed, to allow other things to be GC'd as needed.

Hopefully @benbjohnson will have time to review and get something along these lines merged. :)