@philpearl 2017-05-12
I raised it a year ago tomorrow. I'm no longer interested in merging it, and it was just untidy in my list of PRs

Incidentally, I've moved my application away from boltdb. I ran into the issues with write-amplification warned about in the readme. I've replaced bolt with in-memory structures plus an append only transaction log.